+++ babel.config.js
@@ -1,0 +1,32 @@
+module.exports = api => {
+  const isTest = api.env('test');
+  if (isTest) {
+    return {
+      presets: [
+        [
+          '@babel/env',
+          {
+            targets: {
+              node: 'current'
+            }
+          }
+        ]
+      ]
+    };
+  } else {
+    return {
+      presets: [
+        [
+          '@babel/env',
+          {
+            loose: true,
+            modules: false
+          }
+        ]
+      ],
+      plugins: [
+        // "@babel/plugin-external-helpers"
+      ]
+    };
+  }
+};